+++ src/tools/xml/Field.h
@@ -1414,7 +1414,7 @@ class Enum : public virtual Field
         for (size_t i=0; i < enumType->values.size(); i++)
         {
             if (enumType->values[i] == s)
-                val = i;
+                val = (int)i;
         }
         if (val == -1)
             throw QString("")+"Unknown value for field '"+name+"' of type enum "+enumType->type+": "+s;
